refactor(aigateway): split Gateway into HistoryStore and ToolRunner
